Splitted Missions
* Moved the mission class to a own file.
* Added a folder that contains missions to allow submissions.
* Added install instruction for dependencies.
* Added function to load mission from the missions folder.
