+++ missions/Mission_CartelLeader.yaml
@@ -1,0 +1,7 @@
+title: Kill a cartel leader
+authors: [ Varzeki ]
+attr: STR
+reward: 40
+info: A cartel leader has asked you to help them in an upcoming fight against their rivals!
+success: You fight side by side with the cartel for several hours until none of the rival cartel are left standing.
+fail: Several minutes into the fray, the cartels leader is struck by a stray laser blast, and the leaderless forces scatter along with your chance of payment.
