+++ missions/Mission_JadeFigurine.yaml
@@ -1,0 +1,7 @@
+title: The Jade Figurine
+authors: [ Archina, Undertaker ]
+attr: DEX
+reward: 70
+info: You find a rare jade figurine that belongs to a Chinese politician. He sends a group of ninjas to hunt you down and reclaim it.
+success: Despite all odds, your agile manoeuvring is enough to allude the ninjas. Impressed, they report back to their leader, and you are allowed to keep the figurine.
+fail: Unable to escape the ninjas, you are forced to leave the figurine behind. As you do, the ninjas stop their chase, and you come home to binge on HealBot syringes, licking your wounds on the close escape.
