+++ missions/Mission_Starmap.yaml
@@ -1,0 +1,7 @@
+title: Recover a starmap
+authors: [ Varzeki ]
+attr: DEX
+reward: 60
+info: You are tasked with recovering a starmap from an infested derelict ship!
+success: You accidentally step on a strange spiderlike being on the way out and enrage the hive, only to sprint away with the starmap as the others swarm to eat their wounded.
+fail: The ships security AI mistakes you for one of the infested crew and puts the sector on lockdown. It takes you over an hour and all your ammunition to burn your way through the exit.
